AT&T ksh93 workaround: initialise $! before use
On ksh93, if a background job is launched in a { ... } code block
with a redirection, the $! special parameter (containing the PID of
the background job) does not survive that block. The simplest and
most univeral workaround is to give $! a value before it is used.
See: https://github.com/att/ast/issues/1357

bin/modernish: init stage 2:
- On ksh93, initialise $! by launching a dummy background job.
- Edit comments.