die() overhaul: refactor awk script; remove need for '|| return'
The non-interactive shell version of die() worked fine, but was too
complicated and its awk code was too hard to read. Refactor the awk
code into one single script that terminates the entire program both
from the main shell and from a subshell. The key is that we never
kill the *current* shell process, only exit from it. There should
be no change in die() behaviour for scripts with this commit.

The behaviour of die() on an interactive shell changes as follows:
1. If we die in a subshell, now kill all processes of the current
   job, leaving other jobs alone. If we die in the main interactive
   shell, we do not kill any processes.
2. If we die in the main shell, if SIGINT is ignored and resetting
   the trap and resending SIGINT is ineffective, now just exit. It
   is not worth trying to handle this situation further: it cannot
   be done right, and the current shell situation is broken anyway.
The second change removes the annoying need to add '|| return' to
most die() calls, which was the attempt to handle that situation.

bin/modernish: die():
- Implement changes as above.
- Init: improved detection of POSIX vs legacy 'ps' options.

README.md:
- Document the new behaviour of die() on interactive shells.
- Remove excessive documentation of a corner case with die(); for
  most users, this probably confuses more than it informs.