v0.17.2-dev: new features for sys/cmd/extern
The 'extern' command has gained the following features:

The -p option now does not just search the command in $DEFPATH but
also exports $DEFPATH to it. This is much more desirable and
secure. Take for instance 'extern -p sudo cp ...' -- the reasonable
expectation is that both 'sudo' and 'cp' will be searched in
$DEFPATH. This also applies to the 'sh' and 'cat' forked by 'awk'
when using '| "cat >&2"' to print to standard error. The shell's
builtin 'command -p' fails all this (if it even uses the correct
default path, which it often doesn't).

Assignment-arguments. They are exported to the command. Since
'extern' is a shell function, it cannot use normal preceding
assignments, so these are now provided instead. They are not parsed
after '--' so commands can still contain a '=' (unlike with 'env').

An -u option to unexport a variable to this command only.

lib/modernish/mdl/sys/cmd/extern.mm:
- Regenerate option parser to support option-arguments.
- Add -u option.
- Add assignment-arguments.
- As -p now exports PATH=$DEFPATH, just change PATH directly,
  eliminating the _Msh_p variable.

README.md:
- Update doc.

bin/modernish:
- Version bump.