die(): better awk practice
Accessing an array element in awk creates that element if it
doesn't exist, even if it's something like 'if (foo[bar] == "")'.
While that is harmless in the awk script used by die(), it's not
good practice.

bin/modernish:
- Use '... in ...' in the awk script to check if array elements
  exists instead of comparing array elements to the empty string,
  avoiding the pointless creation of empty array elements.
- Rename parent[] to ppid[] -- better legibility if we use variable
  names corresponding to 'ps' columns.

This commit should not cause any change in the behaviour of die().