QRK_ANDORBG: '&' takes last simple command as bg job, not &&/|| list
On zsh, the '&' operator takes the last simple command
as the background job and not an entire AND-OR list (if any).
In other words,
	a && b || c &
is interpreted as
	a && b || { c & }
and not
	{ a && b || c; } &

This would be a bug, but POSIX (2018 ed.) is ambiguous on the matter.
Ref.: http://austingroupbugs.net/view.php?id=1254
      http://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#tag_18_09_03
      zsh-workers 44705: http://www.zsh.org/mla/workers/2019/msg00741.html

lib/modernish/cap/QRK_ANDORBG.t:
- Added.

lib/modernish/tst/insubshell.t:
- Renamed to lib/modernish/tst/process.t.
- Add regression test for correct QRK_ANDORBG detection.

README.md:
- Document QRK_ANDORBG.