str match: remove broken code handling a dangling final backslash
The code that was supposed to return status 2 if a glob pattern
ended in a final unescaped backslash was broken -- it worked for a
final 'x\' but not for 'x\\\', 'x\\\\\', etc.

Turns out that this is not necessary anyway. With the current match
code, all shells handle this consistently: a dangling final
backslash is implicitly treated as escaped.

This broken check was undocumented, so no documentation adjustment
is necessary.

bin/modernish: str match:
- Remove broken code handling a dangling final backslash.
- Adjust other conditional backslash handling code, removing the
  need for the _Msh_match_backslash variable on initialisation.

lib/modernish/tst/match.t:
- Rewrite regression test for dangling final backslash, now also
  testing an unescaped final backslash following other escaped
  backslashes.