v0.17.5dev: str(): new multi-matching option
One thing that is difficult in shell is matching partial strings
against a set of wholes, with handling of ambiguous matches.
Parsing GNU-style long --options, which can be abbreviated until
they become ambiguous, is a good usage example.

This commit adds a new -M option to str() that makes this sort of
thing easy to do. It works with all binary comparison operators
available in str(), turning them into a multiple matching operators
that compare any number of words against a single pattern.
Ambiguous matches are detected and reflected in the exit status.
Matching words are stored in REPLY separated by newlines.

This commit also makes empty removal handling consistent for all
binary comparison operators. No more different behaviour for
different operators. The usage for all binary operators is now:
	str OPERATOR [ [ WORD ] PATTERN ]
where, if one of the arguments is omitted, the other is assumed to
be the PATTERN, to be matched against an empty WORD.

This commit should only introduce minor incompatibilities with
0.16.x for corner cases related to empty removal handling.

bin/modernish: str():
- Change empty removal handling as described above.
- Add new -M option for multi-matching.
- str isvarname: bugfix: false positive for empty argument.
- Bump version to 0.17.5dev.

lib/modernish/tst/match.t:
- Add multi-matching regression test using str ematch.
- Add regression tests related to empty removal handling.

README.md:
- Overhaul str() documentation.
- Document the new empty removal behaviour.
- Add section documenting the new -M option, with usage example.