sys/cmd/harden: trace(): fix: whitelist exit status 128
trace() is only supposed to die() on system errors (including
signals other than SIGPIPE), not command errors. POSIXly, command
errors should cause an exit with status 125 or less, but 128 is
sometimes used; technically this resembles signal 0, but that's not
actually a signal (some shells use it as the number for the EXIT
pseudosignal for 'trap'). Notably, git uses 128 for many kinds of
command errors.

lib/modernish/mdl/sys/cmd/harden.mm:
- trace(): Don't die on exit status 128 of traced command.