use(): don't use 'command .': may trigger use after free in dash
Sourcing module scripts with 'command .' may trigger a crashing bug
or program flow corruption in dash (all current versions) if a
module then uses 'thisshellhas' to source a feature test that
involves testing an expansion that is invalid on dash (namely,
cap/PSREPLACE.t).

Even if/when this bug is fixed in dash, using 'command .' to stop
the shell exiting on error will unfortunately not be portably
reliable any time soon.

So just use a bare '.' and put up with the shell doing a simple
exit (rather than calling die()) if the file fails to source for
some reason.