var/stack/trap: poptrap() bugfix: clear traps sooner in subshell
Subshells inherit the complete parent shell except for active
non-DIE traps. So all non-DIE traps are cleared when traps are
first modified in the current subshell. This was not handled
correctly by poptrap():

    $ pushtrap 'foo' USR1
    $ (poptrap USR1)
    modernish on dash: poptrap: stack corrupted: USR1

The fix is to call _Msh_clearAllTrapsIfFirstInSubshell() earlier,
before checking whether the trap stack is empty, so that poptrap()
in the subshell correctly finds an empty stack.