sys/base/which: convert relative paths found to absolute
It does not make sense to have 'which' output a relative path; the
path must be usable after changing directories. We now convert all
relative paths, a behaviour that matches GNU 'which'.

The installer had a related bug: when validating the path to a
shell, relative paths were accepted and used as such, so modernish
could be misinstalled with a relative path to the default shell.

lib/modernish/mdl/sys/base/which.mm:
- Reduce excessive intro comments.
- Minor fix: order of options in --help message.
- Shorten subshell warning for 'which -s'.
- Convert any relative path found to an absolute path.
  Also eliminate double slashes if found.

install.sh:
- Bugfix: Use 'which' to verify any user-specified relative shell
  command or path and convert any relative path to absolute.

README.md:
- Document new behaviour.
- Copy-edit the 'which' documentation.