var/local: refactor argument list processing for performance
This commit optimises the performance of processing argument lists
passed to LOCAL blocks after '--'. Instead of shell-quoting each
expanded (and possibly split/globbed) argument and appending it to
a temporary variable, it stores each argument in a temporary array
(on shells with KSHARRAY) or in numbered temporary variables. The
array version comes with an especially noticeable performance gain.

lib/modernish/mdl/var/local.mm:
- Optimise as described above.
- No longer use var/shellquote.