var/shellquote: optimise shellquoteparams
Make the 'shellquoteparams' alias shorter by doing most of the work
in a function.

This requires adding the twice-quoted form of each argument to a
variable so we can 'eval set --' the PPs in one go afterwards
(which removes one level of quoting).

When quoting many PPs (thousands), this is *much* faster than
repeatedly running 'set --' for every parameter. When not quoting
many, it is slower, but not significantly so.