is nonempty: new method for directories
Until now, the directory method for 'is nonempty' read an entire
directory using shell globbing; this is a problem if the directory
has a huge number of files in it. Instead, if we use 'find' to read
just the current directory, we can exec a shell to kill 'find' at
the first result found. The exit status can then be used to
determine if the directory is empty or not.

bin/modernish: is(): nonempty:
- Replace method for testing nonemptiness of directory. This
  depends on the signal number for SIGKILL being 9, and the shell
  adding a multiple of 128 to the signal number to calculate the
  exit status. Both assumptions are de facto portable.

lib/modernish/tst/@sanitychecks.t:
- Add check for the assumption that the shell adds some multiple of
  128 to a signal number to calculate the exit status of a process
  terminated by a signal.