sys/cmd/procsubst: work around bug with 'command'; die on no cmd
ksh93 has a bug where, if 'command' results from an expansion, it
acts like 'command -v' instead of running the command, e.g.:
   set -- command ls; "$@"
acts like 'command -v ls'. This affects modernish process
substitution as it runs the command given as an argument to '%'
using a "$@" expansion. E.g.
   cat < $(% command ls)
shows the path to 'ls' on ksh93. Thus, a workaround is needed.

Bugs with unusual uses of 'command' are fairly common, so to be
future-proof, it seems like a good idea to use this workaround
unconditionally -- also because the cost is negligible.

lib/modernish/mdl/sys/cmd/procsubst.mm:
- To avoid shell bugs, don't allow the "command" command to result
  from the "$@" expansion; shift it out and run "command" directly.
- Die on no command given and on empty command.