aux/fatal.sh: catch FTL_SUBSHTRAP (ksh93 2020); fix FTL_FLOWCORR1
Unfortunately, the latest Korn shell (ksh93) version is too buggy
to run modernish. One of its many bugs is that EXIT traps don't
work when exiting subshells once they've become separate processes.
It is not possible to implement any workaround for this, so it's a
fatal bug. It's a convenient one to catch in the battery of fatal
bug tests, without any extra cost.

lib/modernish/aux/fatal.sh:
- Add FTL_SUBSHTRAP, as per above. Catch this bug by setting a new
  EXIT trap to write the verification string, instead of unsetting
  the EXIT trap and then writing the verification string.
- Correct FTL_FLOWCORR1 detection for zsh:
  . Execution didn't spuriously resume quite at the point I thought
    it did; must take the nested 'case' out of enveloping 'case'.
  . The bug actually exists on zsh < 5.4, not < 5.5. Avoid later
    zsh versions incurring the cost of forking the test subshell.

README.md:
- Document that modernish doesn't run on ksh93 version A 2020.0.0,
  the current "stable" (!) release version.