sys/cmd/readlink: abandon use of nonportable system 'readlink'
It might currently be true that 'readlink -n' is portable on all
systems where 'readlink' is installed, but I can't actually test
every system and be sure; plus, there is no guarantee for the
future. 'readlink' is not a standard command and future 'readlink'
commands could do anything.

By contrast, the ugly way to read a symlink (by using the part
after ' -> ' printed by 'ls -ld') is standardised for all locales,
so guaranteed to keep being portable. It may be unreasonably hard
to make correct and robust, but that's why libraries exist.

lib/modernish/mdl/sys/base/readlink.mm:
- Remove method using system 'readlink'.
- Add sanity checks to the 'ls -ld' method; die on failure.
- Clean up main loop.
- readlink -f:
  * optimisation: since chdir() is now guaranteed to set PWD to a
    physical path, use $PWD instead of a subshell with the 'pwd -P'
    command.
  * optimisation: eliminate a loop iteration for non-symlink path.
  * bugfix: when canonicalising a path ending in '/', don't print
    the final '/'.