sys/base/readlink: make 'readlink -f foo/' act more like GNU
On GNU 'readlink', for a nonexistent path or broken symlink,
'readlink -f nonexist/' is the same as 'readlink -f nonexist'.

This is illogical, because the -f option is documented to allow the
nonexistence of only the last pathname element, and a trailing
slash means the last element is empty. (Indeed, even GNU readlink
outputs nothing and returns 1 for 'readlink -f file/' for an
existing non-directory file.)

However, for compatibility's sake, it's probably best to go along
with this anyway.

-----
One difference remains: for a recursive symlink 'foo' -> 'foo', or
'foo' -> ... -> 'foo', modernish 'readlink -f foo' and 'readlink -f
foo/' both output the canonical path to the symlink, whereas GNU
outputs nothing and returns false. Since a recursive symlink fails
to resolve, so is treated as pointing to a nonexistent file, I
think the modernish behaviour is logical, and no real compatibility
need is served by changing it.