v0.17.14dev: use safe -i: replace fsplit()/glob() with nicer s()/g()
This commit removes the old split/glob manipulation/experimentation
functions for interactive shells in safe mode and replaces them
with s() and g(), precommand modifiers that allow running any
simple command with split and glob applied locally.

Globbing uses the modernish safe failglob or nullglob modes that
prepend './' to expansion results that start with '-' or are equal
to '(' or '!', to stop results being interpreted as options.

The functions are self-documenting (the --help option).

lib/modernish/mdl/safe.mm:
- fsplit(), glob(): Removed.
- s(), g(): Added.

README.md:
- Update brief documentation.

bin/modernish:
- Version bump.