protect use of 'alias'/'unalias' with 'command'
Since 'alias' and 'unalias' are defined as regular builtins,
scripts may define functions under those names. Make sure to use
the real commands in library functions by prefixing 'command'.

On the other hand, in init, outside of library functions, this is
not needed, as init explicitly unaliases and unfunctions these.