BUGFIX: make modernish work on AT&T ksh93 again (re: 591284fb)
Turns out that on AT&T ksh93, 'alias' is a "special builtin" that
exits on error unless 'command' is prefixed. (This is against the
POSIX standard, which specifies 'alias' as a regular builtin.)

Commit 591284fb removed the 'command' prefix from an alias command
that causes an error on ksh93, so modernish no longer initialised
on ksh93.

(Note that version "AJM 93u+ 2012-08-01" is still the only
supported ksh93 version.)

bin/modernish:
- Restore 'command' prefix to the alias command that determines if
  we can alias '[' to a hardened version.