var/loop/find: new -ask primary: simply ask for confirmation
This commit invents a new modernish '-ask' expression primary for
'LOOP find' that asks confirmation of the user. The text of the
prompt may be specified in one optional argument (which cannot
start with '-' or be equal to '!' or '('). Any occurrences of the
characters '{}' within the prompt text are replaced with the
current pathname. If not specified, the default prompt is: '"{}"?'
If the answer is affirmative ('y' or 'Y' in the POSIX locale),
'-ask' yields true, otherwise false. This can be used to make any
part of the expression conditional upon user input.

User input from the shell loop body is not capable of physically
changing find's directory traversal mid-run, but this new primary
is. For example, it can be combined with '-prune' or '-depth n'.

lib/modernish/mdl/var/loop/find.mm:
- Parse new -ask primary, translate to -exec find-ask.sh.

lib/modernish/aux/var/loop/find-ask.sh:
- Add helper script for -ask.

share/doc/modernish/examples/findloop.sh:
- Resurrect & update the old traverse-test example program from
  long-deleted sys/dir/traverse (re: c45f6f44). It demonstrated
  that the directory traversal of 'traverse' could be physically
  influenced from the function/loop body. The new -ask primary now
  provides a way to do this from the 'find' expression instead, so
  this example can work with 'LOOP find'.
- In the comments, document some aspects of 'find' expressions that
  are too intricate for README.md -- particularly that, upon use of
  '-ask' (or '-ok' or '-okdir'), -iterate disables internal results
  grouping optimisation within the current parentheses context in
  order to avoid processing being out of sync with user input.

README.md:
- Document new -ask primary.
- Add a code example demonstrating its use with -prune.