var/loop/find: -ask: dumber & better -iterate optimisation
When -ask, -ok or -okdir are used, internal -iterate grouping
optimisation ('-exec find.sh {} +') is disabled (changed to
'-exec find-ok.sh {} \;') so that loop processing/output won't
go out of sync with user input, i.e. the user immediately sees the
result of each confirmation instead of after 100s of replies.

Until this commit, this -iterate behaviour change was local to the
current set of parentheses. I now believe that's too smart, and
does more harm than good; it's entirely plausible for an -iterate
following later, outside the current parentheses, to be influenced
by the true/false value of -ask/-ok/-okdir within them. This makes
the optimisation undesirable.

For example, this loop would have an optimised -iterate, so the
user would not see any results until after giving 100s of replies:

    LOOP find file in ${1:-$HOME} \
        -type d -and \
            \( -ask 'Traverse directory {}?' -or -prune \) \
        -or -iterate
    DO
        ...
    DONE

To disable the optimisation for interactive use, it would have to
be rewritten as:

    LOOP find file in ${1:-$HOME} \
        -type d -and \
            -not -ask 'Traverse directory {}?' -and -prune \
        -or -iterate
    DO
        ...
    DONE

It seems clear these two should be equivalent, and the first is
easier to read/understand. Therefore, now the optimisation is
disabled for all -iterate primaries that physically follow -ask,
-ok or -okdir in the expression, regardless of parentheses.

If an optimised -iterate is needed in combination with -ask, -ok or
-okdir, the expression should be rewritten so that this -iterate
physically precedes it.

lib/modernish/mdl/var/loop/find.mm:
- Don't push/pop -iterate method when entering/leaving parentheses.

README.md:
- Update code example with -ask to a more intuitive version which
  now no longer fails to set -iterate to the non-optimised
  interactive mode.

share/doc/modernish/examples/findloop.sh:
- Update -iterate optimisation documentation in comments.