bin/modernish: source: simplify init
The modernish version of 'source' was initialised depending on the
behaviour of the shell's 'source' builtin, if any. Unfortunately
the test that is needed to determine if we need the modernish
version involves creating a temporary directory. That might be okay
for a module but it's not okay for the core library.

The modernish version of 'source' will work on all shells, so
the cost of avoiding it on some shells is simply too high.

bin/modernish:
- Unconditionally initialise the modernish version of 'source',
  eliminating complex testing logic involving 'mkdir' in core init.