sys/cmd/source: simplify init; die on not found
The modernish version of 'source' was initialised depending on the
behaviour of the shell's 'source' builtin, if any. Unfortunately
the test that is needed to determine if we need the modernish
version involves creating a temporary directory, which is not ideal.

There is also no guarantee there won't be subtle differences in
current or future 'source' builtin shell implementations that might
cause corner case bugs.

The modernish version of 'source' will work on all shells, so
the cost of avoiding it on some shells is simply too high.

Since all shells will now use the same version if this module is
used, we can also take the opportunity to harden it consistently on
all shells. This makes for consistent behaviour, unlike '.', where
the shell exits on error on most shells but not all.

lib/modernish/mdl/sys/cmd/source.mm:
- Unconditionally initialise the modernish version of 'source',
  eliminating complex testing logic involving 'mkdir' in core init.
- Die on file not found.

README.md:
- Update sys/cmd/source documentation.