install.sh, testshells.sh: don't use readkey
These used modernish 'readkey' to read a single keystroke to answer
a yes/no question. I suppose I was eager to show that off. That did
more harm than good though, as the custom is to expect that the
return key confirms the answer, so users might press it
automatically, thereby accidentally giving the default answer to
the next question (which in install.sh asks where to install it).

install.sh, share/doc/modernish/utils/testshells.sh:
- Use 'read' instead of 'readkey' for y/n questions.
- Some very old broken 'locale' commands output EREs in double
  quotes due to misunderstanding a table in the POSIX spec. I don't
  think we need to take that into account anymore. But add a
  fallback to the POSIX locale {yes,no}expr ERE in any case -- also
  because POSIX doesn't guarantee 'y' or 'n' in non-POSIX locales.