var/loop/find: new --try option: try an expression primary
One problem with `LOOP find` was that, on the one hand, it supports
passing nonstandard expression primaries to the `find` utility so
these extensions can be used in the loop, but on the other hand,
`find` utility selection is automatic so a script can't be sure
which utility is in use.

In a9d93a40 this was somewhat mitigated. This new --try option is
an attempt to further remedy this. It renders "unknown primary"
errors nonfatal: upon encountering a primary that is not supported
by the utility, the loop will quietly abort with exit status 128,
leaving the -name of the invalid primary in the REPLY variable.

This allows a script to test for the presence of certain primaries
and decide what to do.

lib/modernish/mdl/var/loop/find.mm:
- Add --try option.

README.md:
- Document --try option.