var/arith: better arguments check using new ERROR unset readonly
To check for excess arguments with optimal performance, since 2016
the arithmetic shortcuts in var/arith used a rather ugly hack
involving adding an invalid value to the given arithmetic
expression. The resulting error messages were not pretty.

Here is a better way: if we define ERROR as unset and read-only,
we can use ${ERROR:?message here} to trigger an error message. The
variable name is used in the shell's error message, which is why
it's good to call the variable ERROR.

(This could also be ${ERROR?message here} without the colon.
However, on zsh in native mode, readonlies cannot be unset but only
empty. The variant with the extra : works even in that case.)

By nesting this into another parameter substitution, this can be
made conditional upon the existence of a parameter, so for instance
${3+${ERROR:?foo: excess arguments}} exits the shell with the
message "sh: ERROR: foo: excess arguments" if there are 3 or more
positional parameters. This is now used in var/arith.

The word ERROR in capitals might make it look excessively alarming,
but we're not going to make "error" a reserved varname.

bin/modernish:
- Add new ERROR unset readonly on init.

lib/modernish/mdl/var/arith/{cmp,ops}.mm:
- Replace excess arguments checking hack with new use of ERROR,
  yielding proper "excess arguments" error messages on all shells.

README.md:
- Document ERROR.