installer fix: make relative paths work: don't change working dir
Indicating relative paths on the install.sh command line was broken
because the installer changed into the source directory to do the
installing.

The most practical way to fix this (while remaining compatible with
the installer relaunching itself) is to simply not 'cd' before
installing. This required some reworking.

install.sh:
- Don't cd into $MSH_PREFIX.
- Change relative path usage into absolute $MSH_PREFIX/...
- install_file(): Distinguish between a relative path argument
  (install a source file) and an absolute path argument (install a
  bundled script given on the command line).
- Use 'mktemp -t' to create temp dir; it's better practice.
- Temp dir: also cleanup after Ctrl+C (2x -C option to mktemp).
- For patched temp files, use : instead of _ as directory separator
  replacement; _ is used in our filenames so could be ambiguous.

lib/_install/bundle_wrapup.sh:
- Updates related to changes in install.sh.
- When installing bundled scripts, make path absolute first for
  install_file().