install.sh: fix: install files if srcdir is within a hidden .dir
As of the last commit, the full absolute path and not just the
relative path is matched by '-path */.*', which caused nothing to
be installed if one of the components of the parent directory path
starts with '.'. Fixed by using '-name .*' instead.

(Note: in safe mode, there's no need to quote these patterns)