install.sh: allow installing in directories pointed to by symlinks
install.sh would die if, say, $installroot/bin was a symlink to a
directory somewhere else; it was not a directory, so it called
mkdir, which then errored out. So, as of this commit, install.sh
uses 'is -L dir' instead of 'is dir' to check for a directory.

This will confuse the uninstaller, though, as it auto-determines
the install prefix depending on where it finds bin/modernish. So
you have to explicitly indicate it the same installroot with the
'-d' option.