thisshellhas: unbreak --rw=* on mksh <= R50f (FTL_CMDVRESV compat)
Commit 04d6cc90 broke 'thisshellhas --rw=if', etc. on mksh <= R50f
as it introduced an incompatibility with FTL_CMDVRESV. The fix is
simply to use command -V instead of command -v for the initial
non-subshell check, as the output is discarded anyway.

This fix is "forward-ported" from 0.16, where this bug is still
supported as BUG_CMDVRESV, because why not. This way, older mksh
works better when exporting MSH_IGNORE_FATAL_BUGS=y.