bin/modernish: zsh < 5.3: allow 'thisshellhas' to detect 'readonly'
Due to some bug (a race condition?), zsh < 5.3 does not
always include 'readonly' in the output of 'enable', so
that 'thisshellhas --bi=readonly' failed.

Here's a kludge to add it back in for the 0.16 branch, so
'thisshellhas' works as expected.

This can be added to the list of reasons to drop support for
zsh < 5.3 beyond modernish 0.16.x.