examples/dice.sh: remove use of 'tput op'
The terminal code output by 'tput op' is supposed to reset just the
terminal colours, but on some terminals it resets bold as well.
This caused ugly dice.

Eliminating its use also slightly simplifies the code. It just
requires one extra reset for every line of a die.

Also, for the ASCII version, use 'o' instead of '0' for a die dot.