sys/cmd/procsubst: better random FIFO name on all shells
On shells without $RANDOM and on zsh (on which $RANDOM was
ineffective because it is designed to be 100% reproducible), there
were too many 'mkfifo' retries when using process substitution
several times, e.g.:
	diff $(% cmd1) $(% cmd2)
and/or when doing parallel processing.

I've found a way to get a number that's different every time: with
a dummy background job ': &' (which costs a fork, but as it's in
the background it barely costs any time for the invoking shell), a
random-ish PID ends up in "$!". This can be used for the file name.

lib/modernish/mdl/sys/cmd/procsubst.mm:
- Use the "$!" from a dummy bg job in the initial FIFO file name.