sys/base/mktemp: recognise /dev/urandom on QNX
QNX has a non-standard file type, the "named special" (n) which
represents shared memory. /dev/urandom is implemented as a "named
special" on QNX, so that the module's initialisation didn't
recognise it as a charspecial. There is no POSIX command to test
for a named special.

lib/modernish/mdl/sys/base/mktemp.mm:
- Init: only check that /dev/urandom is not a regular file or
  directory before using it. That should be paranoid enough.