thisshellhas: more standards-compliant option/operator parsing
zsh has a builtin command named '-'. Before this commit,
'thisshellhas -' wrongly considered this to be an invalid
short-form shell option name, so exited with status 2.

Also, thisshellhas didn't recognise the '--' end-of-options
delimiter. Its utility is fairly limited here, but it's a standard
thing scripters may expect, so let's add it.

Note that 'thisshellhas --bi=FOO' and 'thisshellhas --rw=FOO'
remain the only way to check for builtins and keywords with
arbitrary names.

bin/modernish:
- Allow checking for builtin or keyword '-' with 'thisshellhas -'.
  This now also allows 'thisshellhas -id' where 'id' is two or more
  characters (thisshellhas -i already checks for a shell option i).
- Recognise the standard '--' end-of-options delimiter, and cause
  it to disable the recognition of special operators starting with
  '-', so that e.g. 'thisshellhas -- -x' checks for a keyword or
  builtin with the name '-x', and 'thisshellhas -- --foo' checks
  for a keyword or builtin with the name '--foo'.

README.md:
- Document the '--' delimiter.
- Correct a documentation error: to check for a builtin/keyword,
  the identifier doesn't actually need to be all-lowercase.