Add VARPREFIX feature detection: ${!prefix@} and ${!prefix*}
The expansions ${!prefix@} and ${!prefix*} yield the names of set
variables whose names begin with 'prefix', in the same way and with
the same quoting effects as $@ and $*, respectively. (bash, ksh93)

The trap stack can use this when it needs to unset all the
trap-related variables when resetting the traps. This improves
performance by avoiding the need to use an external command.

It's also a nice feature for scripts to query and use if available.

lib/modernish/cap/VARPREFIX.t:
- Added.

lib/modernish/cap/BUG_VARPREFIX.t:
- Added. Detect ksh93 bug with VARPREFIX: it finds prefix?* instead
  of prefix* -- vars beginning with 'prefix' but not prefix itself.

lib/modernish/mdl/var/stack/trap.mm:
- If VARPREFIX is detected, use a new version of
  _Msh_clearAllTrapsIfFirstInSubshell() that uses it.
  This avoids the need to invoke 'sed' or fork any subshells.

lib/modernish/tst/var.t:
- Add test for correct VARPREFIX and BUG_VARPREFIX detection.

README.md:
- Document VARPREFIX and BUG_VARPREFIX.