die(): set MSH_HAVE_MERCY variable to kill current subshell only
In certain special cases, die() should not kill the entire program.
The case I can think of is when die() is invoked because a
regression test fails. In that case it is better if it kills the
current subshell and its subprocesses only, so that the parent
process can continue and run the rest of the regression tests.

Note that this is generally a bad idea, as it may allow a script
to keep running in an inconsistent or undefined state. So this
variable should only be set in special cases for a specific reason.

bin/modernish:
- die() interactive version: remove check for _Msh_die_isrunning
  loop check variable (only set in the non-interactive version).
- die() non-interactive version:
  * Unset MSH_HAVE_MERCY on init so it cannot be inherited.
  * If MSH_HAVE_MERCY is set when die() is invoked from a subshell:
    . Consider the current subshell's PID to be the main shell PID
      so that only it and its subprocesses get killed.
    . Don't run DIE traps, as the parent process will continue.
    . Use error msg format "$program: subshell died: $errormsg".

lib/modernish/tst/run.sh:
- Temporarily set MSH_HAVE_MERCY while running each regression test
  so that a failing test cannot kill the entire test suite.

README.md:
- Document the new variable.