Fix 'can write' for directories on shells with '[[' (re: 7e38e735)
Errrm yes, I made a rather stupid logic error in the implementation
of 'can write' on shells that support [[, including bash.

On POSIX shells, the implementation is correct:

| bin/modernish lines 2025 to 2029 in 61b904a
|     ( write )
|             if	test -d "$2"
|             then	test -w "$2" && test -x "$2"
|             else	test -w "$2"
|             fi ;;

Back in 2016, I wrongly thought this was equivalent:

| bin/modernish lines 1823 to 1824 in 61b904a
|     ( write )
|             [[ -d $2 && ( -w $2 && -x $2 ) || -w $2 ]] ;;

The problem is, of course, that if '-d $2' is true and '( -w $2 &&
-x $2 )' is false, then it'll test for '-w $2' instead of returning
false as it should.

Sorry about that. :-/

bin/modernish: can(): write:
- Use correct logic for [[ as we do for the version using test.

Fixes: https://github.com/modernish/modernish/issues/13