tst/sys.t: work around GNU option parsing (re: 88fe72ad)
In Unix, command options are never parsed after the first
non-option argument, but in GNU, they are. This incompatibility
broke the regression test on GNU/Linux systems. POSIXLY_CORRECT=y
would be one way to fix it, but in this case I'll just add a '--'
to disable command option parsing.

Resolves: https://github.com/modernish/modernish/issues/15