QRK_GLOBDOTS: .* harmfully matches '.' and '..'
Until recently, mksh and zsh were the only shells that acted sanely
and refused to match '.' and '..' in a glob pattern like '.*' or
'.?' (imagine accidentally passing '..' to 'cp -R' or 'rm -rf'!).

But AT&T ksh (in version 93u+m) has now fixed that, and bash-5.2
added a new 'globskipdots' option that is enabled by default. With
that, matching '.' and '..' in a glob pattern like '.*' is clearly
no longer the norm, so let's name the quirk and make it detectable.

lib/modernish/cap/QRK_GLOBDOTS.t:
- Added.