FTL_PP_MDLEN => BUG_PP_MDLEN (re: b76a6905)
Unfortunately, dash 0.5.11 < 0.5.11.5 is in fairly widespread use
in the wild, e.g., it's /bin/sh on Debian 11. The fix for this bug
has not been backported, so modernish will have to tolerate it.
Therefore, let's name, document and test it as a non-fatal bug,
allowing modernish to run on these dash versions.