var/stack/trap: Properly fix printing traps (re: 0827e2b2)
The referenced commit still didn't work on zsh, because it stores
trap actions as byte code and reformats them on printing traps. So
'{ _Msh_doTraps ... }' becomes "{$CCn${CCt}_Msh_doTraps ...${CCn}}"
which caused the check for a trap set by modernish to fail.

To fix this properly, initialise a simple "salt" value on loading
the module, made up out of the shell's PID and $RANDOM (or, lacking
that, a subshell's PID). This is added as an extra argument to the
internal trap handler's _Msh_doTraps invocation. This allows
_Msh_doTraps() to use this value as an extra sanity check, but more
importantly, it allows _Msh_printSysTrap() to reliably check for a
trap set by modernish without relying on any whitespace at all.