TRAPPRSUBSH: fix for NetBSD sh (re: 986c2a38)
Current NetBSD sh resets traps for printing purposes in a subshell
when *any* command other than 'trap' is executed within that
subshell, even ':'. This is not a version of TRAPPRSUBSH that we
can use in the var/stack/trap module; what we rely on is that the
shell only resets the traps for printing purposes when a new trap
is set within the subshell. So the feature test incorrectly detects
the feature on NetBSD sh, breaking v=$(trap) for the trap stack.

lib/modernish/cap/TRAPPRSUBSH.t:
- Add a dummy command to avoid detecting the feature on NetBSD sh,
  so that the fallback detection loop for v=$(trap) is used instead
  (see lib/modernish/mdl/var/stack/trap.mm lines 375-383).